skillsakhi_backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
psycopg2-binary==2.9.10
pandas==2.2.3
numpy==2.2.2
scikit-learn==1.6.1
gunicorn==23.0.0
